updated snapshots and failing tests
