+++ src/common/components/Navbar/Navbar.jsx
@@ -1,37 +1,45 @@
-import React, { useState, useRef, useEffect } from "react";
-import { useTranslation } from "react-i18next";
-import { Link, NavLink, useNavigate } from "react-router-dom";
-import { Hub } from "aws-amplify/utils";
-import { useDispatch, useSelector } from "react-redux";
-import {
-  IoPeopleOutline,
-  IoLogInOutline,
-  IoNotificationsOutline,
-} from "react-icons/io5";
-import { BiDonateHeart } from "react-icons/bi";
+import React, { useState, useEffect } from "react";
+import { Drawer, Menu, MenuItem, IconButton } from "@mui/material";
 import LOGO from "../../../assets/logo.svg";
+import { NavLink, useNavigate } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
+import { useTranslation } from "react-i18next";
+
+// Individual imports for outlined icons
+import HomeOutlinedIcon from "@mui/icons-material/HomeOutlined";
+import PeopleOutlinedIcon from "@mui/icons-material/PeopleOutlined";
+import ContactMailOutlinedIcon from "@mui/icons-material/ContactMailOutlined";
+import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
+import Diversity1OutlinedIcon from "@mui/icons-material/Diversity1Outlined";
+import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
+import MenuIcon from "@mui/icons-material/Menu";
+import VolunteerActivismOutlinedIcon from "@mui/icons-material/VolunteerActivismOutlined";
+import Diversity3Icon from "@mui/icons-material/Diversity3";
+import GroupsIcon from "@mui/icons-material/Groups";
+import CrisisAlertIcon from "@mui/icons-material/CrisisAlert";
+import LogoutIcon from "@mui/icons-material/Logout";
+import AccountCircleIcon from "@mui/icons-material/AccountCircle";
+
 import DEFAULT_PROFILE_ICON from "../../../assets/Landingpage_images/ProfileImage.jpg";
-import "./NavBar.css";
+import { IoLogInOutline } from "react-icons/io5";
 import { logout } from "../../../redux/features/authentication/authActions";
-import { GET_NOTIFICATIONS } from "../../../services/requestServices";
-import { useNotifications } from "../../../context/NotificationContext";
-import OurTeamIcon from "../../../assets/Our_Team_SVG.svg";
-import OurMissionIcon from "../../../assets/Our_Mission.svg";
 
 const Navbar = () => {
-  const { t } = useTranslation();
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const [isProfileDropdownOpen, setIsProfileDropdownOpen] = useState(false);
-  const [isLogoutModalOpen, setIsLogoutModalOpen] = useState(false); // Logout modal state
-  const [profileIcon, setProfileIcon] = useState(DEFAULT_PROFILE_ICON);
+  const [anchorEl, setAnchorEl] = useState(null);
+  const [volunteerOpenMenu, setVolunteerOpenMenu] = useState(false);
+  const [aboutUsOpenMenu, setAboutUsOpenMenu] = useState(false);
+  const [profileOpenMenu, setProfileOpenMenu] = useState(false);
+  const [drawerOpen, setDrawerOpen] = useState(false);
   const navigate = useNavigate();
-  const reduxDispatch = useDispatch();
+  const dispatch = useDispatch();
+
+  const [isLogoutModalOpen, setIsLogoutModalOpen] = useState(false);
+  const [profileIcon, setProfileIcon] = useState(DEFAULT_PROFILE_ICON);
   const { user } = useSelector((state) => state.auth);
-  const profileDropdownRef = useRef(null);
+
   const [hasUnsavedChanges, setHasUnsavedChanges] = useState(false);
-  // const [notifications, setNotifications] = useState([]);
-  const { state, dispatch: notificationDispatch } = useNotifications();
-  const [newNotificationCount, setNewNotificationCount] = useState(0);
+
+  const { t } = useTranslation();
 
   useEffect(() => {
     const savedProfilePhoto = localStorage.getItem("profilePhoto");
@@ -58,6 +66,7 @@ const Navbar = () => {
     window.addEventListener("storage", handleStorageChange);
     window.addEventListener("profile-photo-updated", handleProfilePhotoUpdated);
     window.addEventListener("unsaved-changes", handleUnsavedChanges);
+
     return () => {
       window.removeEventListener("storage", handleStorageChange);
       window.removeEventListener(
@@ -68,110 +77,7 @@ const Navbar = () => {
     };
   }, []);
 
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (
-        profileDropdownRef.current &&
-        !profileDropdownRef.current.contains(event.target)
-      ) {
-        setIsProfileDropdownOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-
-    return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, []);
-
-  useEffect(() => {
-    if (!user?.userId) return; // don't fetch unless user is logged in
-
-    const notificationButton = document.getElementById("notificationButton");
-
-    if (notificationButton) {
-      notificationButton.style.display = user ? "flex" : "none";
-    }
-    const fetchNotifications = async () => {
-      try {
-        // const response = await GET_NOTIFICATIONS(); //Call the funciton as of now we are assigning the static data
-        // const rawNotifications = response.data.notifications || [];
-        const rawNotifications = [
-          {
-            type: "Volunteer",
-            title: "New Match Request",
-            message: "You have new Volunteer match request in Logistics",
-            date: "Mar 15, 2023, 10:30 AM",
-          },
-          {
-            type: "Volunteer",
-            title: "New Match Request",
-            message: "Hospital",
-            date: "Jun 15, 2023, 10:30 AM",
-          },
-          {
-            type: "Volunteer",
-            title: "Logistic Help",
-            message: "Logistics",
-            date: "Nov 15, 2023, 10:30 AM",
-          },
-          {
-            type: "helpRequest",
-            title: "Educational Help",
-            message: "Need help with Logistics",
-            date: "Dec 16, 2023, 10:30 AM",
-          },
-          {
-            type: "Volunteer",
-            title: "New Match Request",
-            message: "Education",
-            date: "Jan 15, 2023, 10:30 AM",
-          },
-        ];
-
-        // ✅ Add unique ID using crypto.randomUUID()
-        const notificationsWithIds = rawNotifications.map((note) => ({
-          ...note,
-          id: crypto.randomUUID(),
-        }));
-
-        notificationDispatch({
-          type: "SET_NOTIFICATIONS",
-          payload: notificationsWithIds,
-        });
-        const existing = new Set(
-          state.notifications.map((n) => n.message + n.date),
-        );
-        const newOnes = notificationsWithIds.filter(
-          (n) => !existing.has(n.message + n.date),
-        );
-
-        // Only update if new notifications exist
-        if (newOnes.length > 0) {
-          notificationDispatch({
-            type: "SET_NOTIFICATIONS",
-            payload: [...state.notifications, ...newOnes],
-          });
-
-          setNewNotificationCount((prev) => prev + newOnes.length);
-        }
-      } catch (error) {
-        console.error("Error fetching Notifications:", error);
-      }
-    };
-
-    fetchNotifications(); // Comment it after call ing the funciton below
-
-    const interval = setInterval(
-      "call fetchNotifications() here",
-      2 * 60 * 1000,
-    ); // fetch every 2 min
-
-    return () => clearInterval(interval);
-  }, [notificationDispatch, user]);
-
-  const handleDropdownClick = () => setIsDropdownOpen(!isDropdownOpen);
-
-  const handleLinkClick = (e) => {
+  const handleLinkClick = (e, route) => {
     if (hasUnsavedChanges) {
       e.preventDefault();
       if (
@@ -180,250 +86,447 @@ const Navbar = () => {
         )
       ) {
         setHasUnsavedChanges(false);
-        const targetUrl = e.currentTarget.getAttribute("href");
 
-        if (targetUrl) {
-          navigate(targetUrl);
-        }
+        setVolunteerOpenMenu(false);
+        setAboutUsOpenMenu(false);
+        setProfileOpenMenu(false);
 
-        // Close dropdown if open
-        if (isDropdownOpen) {
-          setIsDropdownOpen(false);
-        }
+        navigate(route);
       } else {
-        // User chose to stay, do nothing
         return;
       }
-    } else if (isDropdownOpen) {
-      setIsDropdownOpen(false);
     }
-  };
+    // Close all dropdowns
+    setVolunteerOpenMenu(false);
+    setAboutUsOpenMenu(false);
+    setProfileOpenMenu(false);
 
-  const handleDropdownItemClick = (e, route) => {
-    if (hasUnsavedChanges) {
-      e.preventDefault();
-      if (
-        window.confirm(
-          "You have unsaved changes. Do you want to proceed without saving?",
-        )
-      ) {
-        setHasUnsavedChanges(false);
-        setIsDropdownOpen(false);
-        navigate(route);
-      }
-    } else {
-      setIsDropdownOpen(false);
-      navigate(route);
-    }
-  };
-
-  const handleNotificationsClick = () => {
-    console.log("notifications");
-  };
-
-  const handleProfileClick = (e) => {
-    if (hasUnsavedChanges) {
-      e.preventDefault();
-      if (
-        window.confirm(
-          "You have unsaved changes. Do you want to proceed without saving?",
-        )
-      ) {
-        setHasUnsavedChanges(false);
-        navigate("/profile");
-        setIsProfileDropdownOpen(false);
-      }
-    } else {
-      navigate("/profile");
-      setIsProfileDropdownOpen(false);
-    }
+    navigate(route);
   };
 
   const handleLogoutClick = (e) => {
-    e.preventDefault();
     if (hasUnsavedChanges) {
+      e.preventDefault();
       if (
         window.confirm(
           "You have unsaved changes. Do you want to proceed without saving?",
         )
       ) {
         setHasUnsavedChanges(false);
+
+        setVolunteerOpenMenu(false);
+        setAboutUsOpenMenu(false);
+        setProfileOpenMenu(false);
+
         setIsLogoutModalOpen(true);
+      } else {
+        return;
       }
     } else {
+      setVolunteerOpenMenu(false);
+      setAboutUsOpenMenu(false);
+      setProfileOpenMenu(false);
+
       setIsLogoutModalOpen(true);
     }
   };
 
   const handleSignOut = () => {
-    reduxDispatch(logout());
+    dispatch(logout());
     setIsLogoutModalOpen(false);
     navigate("/login");
   };
 
-  return (
-    <div className="navbar navbar-sm navbar-gradient-bg">
-      <div className="navbar-start">
-        <Link
-          to="/dashboard"
-          className="text-3xl font-semibold"
-          onClick={handleLinkClick}
-        >
-          <img src={LOGO} alt="logo" className="w-14 h-14" />
-        </Link>
-      </div>
+  const handleDrawerClick = (e, route) => {
+    // Close all dropdowns
+    setVolunteerOpenMenu(false);
+    setAboutUsOpenMenu(false);
 
-      <div className="navbar-end gap-10">
-        <div className="dropdown">
-          <button
-            className={`font-semibold flex flex-col items-center ${isDropdownOpen ? "active" : ""}`}
-            onClick={handleDropdownClick}
-          >
-            <IoPeopleOutline className="mr-1 text-xl" />
-            {t("ABOUT")}
-          </button>
-          <ul
-            tabIndex={0}
-            className={`menu menu-md dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box w-52 ${isDropdownOpen ? "block" : "hidden"}`}
-            onClick={handleLinkClick}
-          >
-            <li>
-              <NavLink to="/directors" name="directors">
-                <img src={OurTeamIcon} alt="Team" className="w-5 h-5" />
-                {t("Our Team")}
-              </NavLink>
-            </li>
-            <li>
-              <NavLink to="/our-mission" name="our-mission">
-                <img src={OurMissionIcon} alt="Mission" className="w-5 h-5" />
-                {t("Our Mission")}
-              </NavLink>
-            </li>
-          </ul>
+    // Close Drawer
+    setDrawerOpen(false);
+
+    navigate(route);
+  };
+
+  const handleVSMenuClick = (event) => {
+    setAnchorEl(event.currentTarget);
+    setVolunteerOpenMenu(true);
+  };
+
+  const handleVSMenuClose = () => {
+    setAnchorEl(null);
+    setVolunteerOpenMenu(false);
+  };
+
+  const handleAUMenuClick = (event) => {
+    setAnchorEl(event.currentTarget);
+    setAboutUsOpenMenu(true);
+  };
+
+  const handleAUMenuClose = () => {
+    setAnchorEl(null);
+    setAboutUsOpenMenu(false);
+  };
+
+  const handlePMenuClick = (event) => {
+    setAnchorEl(event.currentTarget);
+    setProfileOpenMenu(true);
+  };
+
+  const handlePMenuClose = () => {
+    setAnchorEl(null);
+    setProfileOpenMenu(false);
+  };
+
+  const toggleDrawer = (open) => {
+    setDrawerOpen(open);
+  };
+
+  return (
+    <nav className="bg-[#FFFFFF] sticky top-0 h-[113px] w-full z-10 shadow-m font-josefin p-4">
+      <div className="flex items-center justify-between w-full px-4 h-full">
+        {/* Logo aligned to the left */}
+        <div
+          className="flex items-center cursor-pointer"
+          onClick={(e) => handleLinkClick(e, "/")}
+        >
+          <img src={LOGO} alt="Company Logo" className="w-[60px] h-[60px]" />
         </div>
 
-        <NavLink
-          to="/donate"
-          className="font-semibold flex flex-col items-center"
-          onClick={handleLinkClick}
-        >
-          <BiDonateHeart className="mr-1 text-xl" />
-          {t("DONATE")}
-        </NavLink>
-
-        <NavLink
-          to="/notifications"
-          id="notificationButton"
-          className="font-semibold flex flex-col items-center"
-          onClick={() => {
-            handleNotificationsClick;
-            setNewNotificationCount(0); // ✅ Reset count when clicked
-          }}
-        >
+        {/* Desktop Menu (visible only on larger screens) */}
+        <div className="hidden md:flex items-center space-x-8 ml-auto">
           <div className="relative">
-            <IoNotificationsOutline className="mr-1 text-xl" />
-            {newNotificationCount > 0 && (
-              <span className="absolute -top-1 right-0 bg-red-600 text-white text-xs px-1 rounded-full">
-                {newNotificationCount}
-              </span>
+            <button
+              onClick={(e) => handleLinkClick(e, "/")}
+              // className="text-black flex items-center hover:text-gray-600 text-base"
+              className="text-black hover:text-gray-600 flex items-center text-base"
+            >
+              <HomeOutlinedIcon className="mr-2" /> {t("Home")}
+            </button>
+          </div>
+
+          {/* About Us Dropdown */}
+          <div className="relative">
+            <button
+              onClick={handleAUMenuClick}
+              className="text-black flex items-center hover:text-gray-600 text-base"
+            >
+              <PeopleOutlinedIcon className="mr-2" /> {t("About Us")}
+              <ArrowDropDownIcon />
+            </button>
+            {aboutUsOpenMenu && (
+              <Menu
+                anchorEl={anchorEl}
+                open={aboutUsOpenMenu}
+                onClose={handleAUMenuClose}
+                PaperProps={{
+                  style: {
+                    position: "absolute",
+                    right: 0,
+                    top: "40px",
+                    zIndex: 1300, // to appear above other elements
+                    maxWidth: "fit-content",
+                  },
+                }}
+              >
+                <MenuItem onClick={(e) => handleLinkClick(e, "/directors")}>
+                  <GroupsIcon className="mr-2" /> {t("Our Team")}
+                </MenuItem>
+                <MenuItem onClick={(e) => handleLinkClick(e, "/vision")}>
+                  <CrisisAlertIcon className="mr-2" /> {t("Our Vision")}
+                </MenuItem>
+              </Menu>
             )}
           </div>
-          {t("NOTIFICATIONS")}
-        </NavLink>
-        {/* <button
-          className="font-semibold flex flex-col items-center"
-          id="notificationButton"
-          onClick={handleNotificationsClick}
-        >
-          <IoNotificationsOutline className="mr-1 text-xl" />
-          {t("NOTIFICATIONS")}
-        </button> */}
 
+          {/* Volunteer Services Dropdown */}
+          <div className="relative">
+            <button
+              onClick={handleVSMenuClick}
+              className="text-black flex items-center hover:text-gray-600 text-base"
+            >
+              <Diversity1OutlinedIcon className="mr-2" />
+              {t("Volunteer Services")}
+              <ArrowDropDownIcon />
+            </button>
+            {volunteerOpenMenu && (
+              <Menu
+                anchorEl={anchorEl}
+                open={volunteerOpenMenu}
+                onClose={handleVSMenuClose}
+                PaperProps={{
+                  style: {
+                    position: "absolute",
+                    right: 0,
+                    top: "40px",
+                    zIndex: 1300, // to appear above other elements
+                    maxWidth: "fit-content",
+                  },
+                }}
+              >
+                <MenuItem
+                  onClick={(e) => handleLinkClick(e, "/how-we-operate")}
+                >
+                  <GroupsIcon className="mr-2" /> {t("How We Operate")}
+                </MenuItem>
+                <MenuItem
+                  onClick={(e) => handleLinkClick(e, "/how-we-operate")}
+                >
+                  <Diversity3Icon className="mr-2" /> {t("Our Collaborators")}
+                </MenuItem>
+              </Menu>
+            )}
+          </div>
+
+          <div className="relative">
+            <button
+              onClick={(e) => handleLinkClick(e, "/contact")}
+              // className="text-black flex items-center hover:text-gray-600 text-base"
+              className="text-black hover:text-gray-600 flex items-center text-base"
+            >
+              <ContactMailOutlinedIcon className="mr-2" /> {t("Contact Us")}
+            </button>
+          </div>
+        </div>
+
+        {/* Donate button aligned to the rightmost */}
+
+        <div className="ml-auto mr-3">
+          <button
+            onClick={(e) => handleLinkClick(e, "/donate")}
+            // className="text-black flex items-center hover:text-gray-600 text-base"
+            className="bg-blue-500 text-white py-2 px-6 rounded-full hover:bg-blue-600 ml-auto flex items-center"
+          >
+            <VolunteerActivismOutlinedIcon className="mr-2 text-base" />
+            {t("Donate")}
+          </button>
+        </div>
+
+        {/* Login Part */}
         {user?.userId ? (
           <div
-            className="relative flex flex-col items-center"
-            ref={profileDropdownRef}
+            className="relative flex flex-col items-center mr-2"
+            // ref={profileDropdownRef}
           >
             <div className="flex items-center">
-              <img
-                src={profileIcon}
-                alt="Profile Icon"
-                className="w-8 h-8 rounded-full cursor-pointer"
-                onClick={() => {
-                  if (hasUnsavedChanges) {
-                    if (
-                      window.confirm(
-                        "You have unsaved changes. Do you want to proceed without saving?",
-                      )
-                    ) {
-                      setHasUnsavedChanges(false);
-                      setIsProfileDropdownOpen(!isProfileDropdownOpen);
+              <IconButton color="inherit" edge="end">
+                <img
+                  src={profileIcon}
+                  alt="Profile Icon"
+                  className="w-8 h-8 rounded-full cursor-pointer"
+                  onClick={(e) => {
+                    if (hasUnsavedChanges) {
+                      if (
+                        window.confirm(
+                          "You have unsaved changes. Do you want to proceed without saving?",
+                        )
+                      ) {
+                        setHasUnsavedChanges(false);
+                        handlePMenuClick(e);
+                      }
+                    } else {
+                      handlePMenuClick(e);
                     }
-                  } else {
-                    setIsProfileDropdownOpen(!isProfileDropdownOpen);
-                  }
-                }}
-              />
+                  }}
+                />
+              </IconButton>
             </div>
-            {isProfileDropdownOpen && (
-              <ul className="profile-dropdown">
-                <li
-                  className="p-2 hover:bg-gray-100 cursor-pointer"
-                  onClick={handleProfileClick}
-                >
-                  {t("PROFILE")}
-                </li>
-                <li
-                  className="p-2 hover:bg-gray-100 cursor-pointer"
-                  onClick={handleLogoutClick}
-                >
-                  {t("LOGOUT")}
-                </li>
-              </ul>
+            {profileOpenMenu && (
+              <Menu
+                anchorEl={anchorEl}
+                open={profileOpenMenu}
+                onClose={handlePMenuClose}
+                PaperProps={{
+                  style: {
+                    position: "absolute",
+                    right: 0,
+                    top: "40px",
+                    zIndex: 1300, // to appear above other elements
+                    maxWidth: "fit-content",
+                  },
+                }}
+              >
+                <MenuItem onClick={(e) => handleLinkClick(e, "/profile")}>
+                  <AccountCircleIcon className="mr-2" />
+                  {t("Profile")}
+                </MenuItem>
+                <MenuItem onClick={(e) => handleLogoutClick()}>
+                  <LogoutIcon className="mr-2" />
+                  {t("Logout")}
+                </MenuItem>
+              </Menu>
             )}
           </div>
         ) : (
           <NavLink
             to="/login"
-            className="font-semibold flex flex-col items-center"
+            className="font-semibold flex flex-col items-center mr-2"
             id="loginButton"
-            onClick={handleLinkClick}
+            onClick={(e) => handleLinkClick(e, "/login")}
           >
             <IoLogInOutline className="mr-1 text-xl" />
             {t("LOGIN")}
           </NavLink>
         )}
-      </div>
 
-      {/* Logout Confirmation Modal */}
-      {isLogoutModalOpen && (
-        <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50">
-          <div className="bg-white p-6 rounded-lg shadow-lg">
-            <p className="text-lg font-semibold mb-4">
-              {t(
-                "Are you sure, you want to log out? There may be unsaved data on the page.",
-              )}
-            </p>
-            <div className="mt-8 flex justify-end gap-2">
-              <button
-                type="submit"
-                className="py-2 px-4 bg-blue-500 text-white rounded-md mr-2 hover:bg-blue-600"
-                onClick={handleSignOut}
-              >
-                {t("Logout")}
-              </button>
-              <button
-                onClick={() => setIsLogoutModalOpen(false)}
-                type="button"
-                className="py-2 px-4 bg-gray-500 text-white rounded-md hover:bg-gray-600"
-              >
-                {t("Cancel")}
-              </button>
+        {/* Logout Confirmation Modal */}
+        {isLogoutModalOpen && (
+          <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50">
+            <div className="bg-white p-6 rounded-lg shadow-lg">
+              <p className="text-lg font-semibold mb-4">
+                {
+                  "Are you sure, you want to log out? There may be unsaved data on the page."
+                }
+              </p>
+              <div className="mt-8 flex justify-end gap-2">
+                <button
+                  type="submit"
+                  className="py-2 px-4 bg-blue-500 text-white rounded-md mr-2 hover:bg-blue-600"
+                  onClick={handleSignOut}
+                >
+                  {t("Logout")}
+                </button>
+                <button
+                  onClick={() => setIsLogoutModalOpen(false)}
+                  type="button"
+                  className="py-2 px-4 bg-gray-500 text-white rounded-md hover:bg-gray-600"
+                >
+                  {t("Cancel")}
+                </button>
+              </div>
             </div>
           </div>
+        )}
+
+        {/* Mobile Menu Icon (visible only on mobile) */}
+        <div className="md:hidden mr-4">
+          <IconButton
+            color="inherit"
+            edge="end"
+            onClick={() => toggleDrawer(true)}
+            sx={{ display: { xs: "block", sm: "none" } }}
+          >
+            <MenuIcon />
+          </IconButton>
         </div>
-      )}
-    </div>
+      </div>
+
+      {/* Mobile Drawer (Replaces Navbar on Mobile) */}
+      <Drawer
+        anchor="top"
+        open={drawerOpen}
+        onClose={() => toggleDrawer(false)}
+        variant="temporary"
+        sx={{
+          "& .MuiDrawer-paper": {
+            top: "113px", // Adjusting the top offset so it aligns with the navbar height
+            width: "100%",
+            paddingTop: "0",
+            height: "auto",
+          },
+        }}
+      >
+        <div className="p-6 w-64">
+          <div className="block text-black py-2 flex items-center">
+            <button
+              onClick={(e) => handleDrawerClick(e, "/")}
+              className="text-black flex items-center hover:text-blue-600"
+            >
+              <HomeOutlinedIcon className="mr-2" /> {t("Home")}
+            </button>
+          </div>
+
+          <div className="block text-black py-2 flex items-center">
+            <button
+              onClick={handleAUMenuClick}
+              className="text-black flex items-center hover:text-blue-600"
+            >
+              <PeopleOutlinedIcon className="mr-2" /> {t("About Us")}
+              <ArrowDropDownIcon />
+            </button>
+            {aboutUsOpenMenu && (
+              <Menu
+                anchorEl={anchorEl}
+                open={aboutUsOpenMenu}
+                onClose={handleAUMenuClose}
+                PaperProps={{
+                  style: {
+                    position: "absolute",
+                    right: 0,
+                    top: "40px",
+                    zIndex: 1300, // to appear above other elements
+                    maxWidth: "fit-content",
+                  },
+                }}
+              >
+                <MenuItem onClick={(e) => handleDrawerClick(e, "/directors")}>
+                  <GroupsIcon className="mr-2" /> {t("Our Team")}
+                </MenuItem>
+                <MenuItem onClick={(e) => handleDrawerClick(e, "/vision")}>
+                  <CrisisAlertIcon className="mr-2" /> {t("Our Vision")}
+                </MenuItem>
+              </Menu>
+            )}
+          </div>
+
+          <div className="block text-black py-2 flex items-center">
+            <button
+              onClick={handleVSMenuClick}
+              className="text-black flex items-center hover:text-blue-600"
+            >
+              <Diversity1OutlinedIcon className="mr-2" />
+              {t("Volunteer Services")}
+              <ArrowDropDownIcon />
+            </button>
+            {volunteerOpenMenu && (
+              <Menu
+                anchorEl={anchorEl}
+                open={volunteerOpenMenu}
+                onClose={handleVSMenuClose}
+                PaperProps={{
+                  style: {
+                    position: "absolute",
+                    right: 0,
+                    top: "40px",
+                    zIndex: 1300, // to appear above other elements
+                    maxWidth: "fit-content",
+                  },
+                }}
+              >
+                <MenuItem
+                  onClick={(e) => handleDrawerClick(e, "/how-we-operate")}
+                >
+                  <GroupsIcon className="mr-2" /> {t("How We Operate")}
+                </MenuItem>
+                <MenuItem
+                  onClick={(e) => handleDrawerClick(e, "/how-we-operate")}
+                >
+                  <Diversity3Icon className="mr-2" /> {t("Our Collaborators")}
+                </MenuItem>
+              </Menu>
+            )}
+          </div>
+
+          <div className="block text-black py-2 flex items-center">
+            <button
+              onClick={(e) => handleDrawerClick(e, "/contact")}
+              className="text-black flex items-center hover:text-blue-600"
+            >
+              <ContactMailOutlinedIcon className="mr-2" /> {t("Contact Us")}
+            </button>
+          </div>
+
+          <div className="block text-black py-2 flex items-center">
+            <button
+              onClick={(e) => handleDrawerClick(e, "/donate")}
+              className="text-black flex items-center hover:text-blue-600"
+            >
+              <FavoriteBorderIcon className="mr-2" /> {t("Donate")}
+            </button>
+          </div>
+        </div>
+      </Drawer>
+    </nav>
   );
 };
